+++ packages/kofono/CHANGELOG.md
@@ -13,6 +13,7 @@
 - rework custom extensions support
 - transform DataTree class methods into three functions
 - remove deprecated form logs
+- fix some array edge cases with DataSelector
 
 # 0.7.0 - 2026-03-29